RUMM-149 Move LogFileStrategy in domain package
